requirements.txt
requests==2.31.0
types-requests==2.31.0.2
pandas==2.1.0
matplotlib==3.8.0
geopy==2.4.0
pyserial==3.5